rename: componentsフォルダ名を修正
 - componentsフォルダの名称にスペルミスがあったため、正しいスペルに修正しました
