+++ frontend/app/components/elements/footer/footer.tsx
